+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/reactions/user/internal/SingleReactionView.kt
@@ -23,14 +23,18 @@ import android.widget.FrameLayout
 import io.getstream.chat.android.ui.common.extensions.internal.createStreamThemeWrapper
 import io.getstream.chat.android.ui.common.extensions.internal.streamThemeInflater
 import io.getstream.chat.android.ui.databinding.StreamUiItemMessageReactionBinding
-import io.getstream.chat.android.ui.message.list.reactions.view.ViewReactionsViewStyle
-import io.getstream.chat.android.ui.message.list.reactions.view.internal.ViewReactionsBubbleDrawer
+import io.getstream.chat.android.ui.message.list.reactions.user.SingleReactionViewStyle
+import io.getstream.chat.android.ui.message.list.reactions.view.MessageOptionsUserReactionOrientation
+import io.getstream.chat.android.ui.message.list.reactions.view.getUserReactionOrientation
+import io.getstream.chat.android.ui.message.list.reactions.view.isOrientedTowardsStart
 
 internal class SingleReactionView : FrameLayout {
     private val binding = StreamUiItemMessageReactionBinding.inflate(streamThemeInflater, this, true)
-    private lateinit var reactionsViewStyle: ViewReactionsViewStyle
-    private lateinit var bubbleDrawer: ViewReactionsBubbleDrawer
+    private lateinit var reactionsViewStyle: SingleReactionViewStyle
+    private lateinit var bubbleDrawer: SingleReactionViewBubbleDrawer
     private var isMyMessage: Boolean = false
+    private val messageOrientation: MessageOptionsUserReactionOrientation
+        get() = reactionsViewStyle.reactionOrientation.getUserReactionOrientation()
 
     constructor(context: Context) : super(context.createStreamThemeWrapper()) {
         init(context, null)
@@ -51,26 +55,28 @@ internal class SingleReactionView : FrameLayout {
     fun setReaction(userReactionItem: UserReactionItem) {
         // according to the design, current user reactions have the same style
         // as reactions on the current user messages in the message list
-        this.isMyMessage = !userReactionItem.isMine
+        this.isMyMessage = userReactionItem.isMine
         binding.reactionIcon.setImageDrawable(userReactionItem.drawable)
         invalidate()
     }
 
     override fun onDraw(canvas: Canvas) {
         super.onDraw(canvas)
+
+        val isOrientedTowardsStart = messageOrientation.isOrientedTowardsStart(isMyMessage)
+
         bubbleDrawer.drawReactionsBubble(
-            context,
-            canvas,
-            width,
+            context = context,
+            canvas = canvas,
+            bubbleWidth = width,
             isMyMessage = isMyMessage,
-            isSingleReaction = true,
-            inverseBubbleStyle = true
+            isOrientedTowardsStart = isOrientedTowardsStart
         )
     }
 
     private fun init(context: Context, attrs: AttributeSet?) {
-        this.reactionsViewStyle = ViewReactionsViewStyle(context, attrs)
-        this.bubbleDrawer = ViewReactionsBubbleDrawer(reactionsViewStyle)
+        this.reactionsViewStyle = SingleReactionViewStyle(context, attrs)
+        this.bubbleDrawer = SingleReactionViewBubbleDrawer(reactionsViewStyle)
 
         setWillNotDraw(false)
         minimumHeight = reactionsViewStyle.totalHeight